feat/reordered app.js and added unknonw endpoint
