+++ server.js
@@ -1,7 +1,7 @@
 const express = require("express")
 const PORT = 5000;
 
-const { app } = require('./midlewares')
+const { app,unknownEndpoint  } = require('./midlewares/app')
 const db = require('./db/connection')
 
 
@@ -9,6 +9,7 @@ const db = require('./db/connection')
 app.get('/', (req, res)=>{
     res.send(200,{message: 'Bienvenido a Api Task'})
 })
+app.use(unknownEndpoint)
 
 app.listen(PORT, () => { console.log(`Server is running... in port ${PORT}`) });
